+++ blackbox-logging-04/Tracer/Tracer.cpp
@@ -51,7 +51,7 @@ namespace Gbp { namespace Tra {
 			if(m < 1024) m = 1024;
 			
 			// power of two buffers to help with alignment
-			size_t n = 1;
+			int n = 1;
 			while(n < m) n *= 2;
 
 			N += n;
@@ -97,7 +97,7 @@ namespace Gbp { namespace Tra {
 				return false;
 			}
 
-			fprintf(tfile.f(), "Tracer BufferSize=%u DefnCount=%u\n", size_ / sizeof(Slot_t), traceSpecVector_.size());
+			fprintf(tfile.f(), "Tracer BufferSize=%zu DefnCount=%zu\n", size_ / sizeof(Slot_t), traceSpecVector_.size());
 			for(TraceSpecVector::iterator i = traceSpecVector_.begin(); i != traceSpecVector_.end(); ++i)
 			{
 				bool r = (*i)->pTraceDefn->save(tfile.f());
@@ -115,7 +115,7 @@ namespace Gbp { namespace Tra {
 			return false;
 		}
 
-		int r = fwrite(&buffer_[0], 1, buffer_.size(), bfile.f());
+		size_t r = fwrite(&buffer_[0], 1, buffer_.size(), bfile.f());
 		return r == buffer_.size();
 	}
